Delay PatchProc() to give GGST enough time to start up
In rare cases EnumProcessModules() gets called while GGST
is still starting up, causing it to fail.

If the module list in the target process is corrupted or not yet initialized, or if the module list changes during the function call as a result of DLLs being loaded or unloaded, EnumProcessModules may fail or return incorrect information.