Add automatic offset search (#24)
* Add automatic offset search

Searches program memory for the API url and if found sets the offset to the found address. If not found, continues with the predefined offset address. Currently assumes the first instance of the API url found in memory is the one that needs to be overwritten.

* Move to new function

Corrected a few of the comments:
- Moved to independent function
- fixed type issues
- Changed order of API offset checking

* Add searching in chunks

This commit addresses the following comments:
- Memory size of the search buffer.  Its probably a bit small at the moment, but still runs pretty quick.
- Handles the case of if the API url is on region borders. This is pretty inefficient (could be much more efficient if made the chunk size bigger), but since this is the unusual case, its probably fine. Still runs pretty quick.
 - Added in 'already patched' checking. Previously it would cause issues if launched when GG was already running

* Change types to minimizing casting

Change several types from uint32 to uintptr.
rearranged error structure.
removed irrelevant comment